Restructure project layout and update imports
