+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "florence-2-finetuning"
+version = "0.1.0"
+description = "Fine-tuning tools for Microsoft Florence-2"
+authors = [{name = "OpenAI"}]
+
+[tool.setuptools.package-dir]
+"" = "src"
+
+[tool.setuptools.packages.find]
+where = ["src"]
