+++ scripts/generate.py
